feat: refactor install/uninstall function
